Update values.yaml defaults to match real default values
Conventionally, commented out values in the values.yaml file should match the default values for the commented out parameters. Some of the parameters here didn't match, so I've corrected them, as per the documented defaults in [the FAQ](https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/FAQ.md).

It's likely worth discussing whether this section should either be supplemented to include all of the defaults, or whether it should just link to the documentation of the flags in the FAQ. However for now, the defaults listed match their real default values.

Bump version

Add note documenting where to find full list of CA parameters

This adds a note that sends users to the full list of cluster autoscaler parameters in the readme. This is preferable to adding the full list with defaults here, because then we'd have to keep both places in sync with each other.

Update README.md to match values.yaml comments